requirements.txt
pyyaml
jsonschema
pandas
numpy
matplotlib
scipy
argparse